Move CarouselArrow component location to be exported correctly
