+++ source/components/carousel/__tests__/Carousel-test.js
@@ -1,5 +1,4 @@
 import Carousel from '..'
-import Arrow from '../arrow'
 
 describe('Carousel', () => {
   it('should render a simple carousel', () => {
@@ -23,16 +22,5 @@ describe('Carousel', () => {
 
     expect(wrapper.find('.slick-prev')).to.have.length(1)
     expect(wrapper.find('.slick-next')).to.have.length(1)
-  })
-})
-
-describe('Arrow', () => {
-  it('should render a simple carousel arrow', () => {
-    const wrapper = mount(
-      <Arrow direction='prev' />
-    )
-
-    expect(wrapper.find('button').prop('aria-label')).to.eql('Previous')
-    expect(wrapper.find('svg')).to.have.length(1)
   })
 })
